Separate damaged input from unsupported input
- Add waxerr.CodeMalformedInput, exit class 9, HTTP 422.
- Re-sort every demuxer and decoder refusal onto the right one.
- Give container.Warning a Kind; strict escalates damage only.
- Split Info.Notes from Info.Warnings through probe, API and CLI.
- Add Track.SamplesAdvisory; a timeline refuses such a member at plan time.
- Stop capping a decode at a rounded total.
- Upgrade waxlabel to v1.7.0 and retire the rate request.
- Type the MP4 covr atom by MIME, then by bytes.
- Read QuickTime version 2 sound entries by their own layout.
- Bound the WAVEFORMATEX 32-bit fields before they become an int.
- Fix the 32-bit build; add make test-386 and its CI job.

+++ container/riff/mux.go
@@ -90,7 +90,9 @@ func (m *Muxer) Begin(tracks []container.Track) error {
 	if err := t.Fmt.Valid(); err != nil {
 		return err
 	}
-	if t.Fmt.Rate > size32Unknown {
+	// Compared as an int64: size32Unknown is 2^32-1 and Rate is an int, which
+	// is 32 bits on a 32-bit build, where the untyped constant does not fit.
+	if int64(t.Fmt.Rate) > size32Unknown {
 		return waxerr.New(waxerr.CodeUnsupportedFormat, fmt.Sprintf("wav: rate %d does not fit WAV", t.Fmt.Rate))
 	}
 	m.cfg = cfg
